🔧 (install_apt_packages.sh): add user prompt before installing packages 🔧 (install_usb_hdmi_drivers.sh): add user prompt before installing drivers 🔧 (sync_dotfiles_with_system.sh): add user prompt before syncing dotfiles

+++ setup/scripts/install_usb_hdmi_drivers.sh
@@ -1,8 +1,7 @@
-info "Installing DisplayLink USB HDMI drivers..."
-
-git clone https://github.com/AdnanHodzic/displaylink-debian.git
-cd displaylink-debian
-sudo ./displaylink-debian.sh
-rm -rf displaylink-debian
-
-success "DisplayLink USB HDMI drivers installed!"
+if ask "Do you want to install DisplayLink USB HDMI drivers?"; then
+  git clone https://github.com/AdnanHodzic/displaylink-debian.git
+  cd displaylink-debian
+  sudo ./displaylink-debian.sh
+  rm -rf displaylink-debian
+  success "DisplayLink USB HDMI drivers installed!"
+fi
